+++ Example/SwiftyMenu/AppDelegate.swift
@@ -16,6 +16,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
+
+        // For testing RTL support
+//        UIView.appearance().semanticContentAttribute = .forceRightToLeft
+
         return true
     }